+++ back_end/src/main/java/com/example/back_end/controller/AuthController.java
@@ -6,13 +6,10 @@ import com.example.back_end.dto.request.UserCreationRequest;
 import com.example.back_end.dto.response.ApiResponse;
 import com.example.back_end.dto.response.AuthenticationResponse;
 import com.example.back_end.dto.response.IntrospectResponse;
-import com.example.back_end.entity.User;
 import com.example.back_end.service.AuthService;
-import com.example.back_end.service.UserService;
 import com.example.back_end.service.user.UserService;
 import com.nimbusds.jose.JOSEException;
 import jakarta.servlet.http.HttpServletResponse;
-import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
